fix(simpleTable): simpleTableRow child and fixed truncating
