+++ src/components/filter/FilterWrapper.module.scss
@@ -70,7 +70,7 @@ $border-right-color: var(--amino-filter-wrapper-border-right-color);
 }
 
 .dropdownWrapper {
-  z-index: 1;
+  z-index: 5;
   display: flex;
   min-width: 400px;
   position: absolute;
