+++ src/components/simple-table/SimpleTable.module.scss.d.ts
@@ -1,15 +1,9 @@
-export declare const allowTextWrap: string;
 export declare const bordered: string;
-export declare const clickable: string;
-export declare const collapsed: string;
 export declare const collapsible: string;
-export declare const hasContent: string;
 export declare const loading: string;
 export declare const noHeaders: string;
 export declare const noPadding: string;
-export declare const shouldTruncate: string;
 export declare const skeletonCellWrapper: string;
 export declare const styledCheckbox: string;
 export declare const tableContainer: string;
 export declare const tableStyled: string;
-export declare const withHover: string;
