+++ src/components/simple-table/SimpleTable.tsx
@@ -1,16 +1,15 @@
-import React, { Fragment, type ReactNode, useRef } from 'react';
+import React, { type ReactNode } from 'react';
 
 import clsx from 'clsx';
 
 import { Checkbox } from 'src/components/checkbox/Checkbox';
 import { Skeleton } from 'src/components/skeleton/Skeleton';
-import { TableRowCollapse } from 'src/components/table/TableRowCollapse';
 import { Text } from 'src/components/text/Text';
-import { Tooltip } from 'src/components/tooltip/Tooltip';
 import type { BaseProps } from 'src/types/BaseProps';
 import type { ReactComponent } from 'src/types/ReactComponent';
 
 import styles from './SimpleTable.module.scss';
+import { SimpleTableRow } from './SimpleTableRow';
 
 const getFlexAlignment = (alignment: SimpleTableHeaderBaseProps['align']) => {
   switch (alignment) {
@@ -24,23 +23,7 @@ const getFlexAlignment = (alignment: SimpleTableHeaderBaseProps['align']) => {
   }
 };
 
-const getTooltipPlacement = (
-  alignment: SimpleTableHeaderBaseProps['align'],
-) => {
-  switch (alignment) {
-    case 'center':
-      return 'bottom';
-    case 'end':
-      return 'bottom-end';
-    case 'start':
-    default:
-      return 'bottom-start';
-  }
-};
-
-const tooltipDelay = 800;
-
-type SimpleTableHeaderBaseProps = {
+export type SimpleTableHeaderBaseProps = {
   /**
    * Text alignment for a column
    * @default 'start'
@@ -220,90 +203,6 @@ export const SimpleTable = <T extends object>({
   },
   style,
 }: SimpleTableProps<T>) => {
-  const renderHeader = (header: SimpleTableHeader<T>, item: T) => {
-    const value = item[header.key];
-
-    const renderContent = (content: ReactNode) => {
-      const hasRowHoverShowChild = React.Children.toArray(content).some(
-        child => {
-          if (React.isValidElement(child)) {
-            return /row-hover-show|cell-hover-show/.test(child.props.className);
-          }
-          return false;
-        },
-      );
-
-      // If the cell has a child that should only show on hover, don't allow truncating
-      const tdClassNames = clsx(
-        header.textWrapMethod === 'truncate' &&
-          !hasRowHoverShowChild &&
-          styles.shouldTruncate,
-      );
-
-      const containerStyle = {
-        '--amino-cell-min-width': `${header.minWidth || 0}px`,
-      };
-
-      const cellClassNames = clsx(
-        header.noPadding && styles.noPadding,
-        header.textWrapMethod === 'normal' && styles.allowTextWrap,
-      );
-
-      const cellStyle = {
-        textAlign: header.align || 'start',
-      };
-
-      const tooltipPlacement = getTooltipPlacement(header.align);
-
-      if (getRowLink && !selectable.anySelected && !header.disabledLink) {
-        const LinkComponent = CustomLinkComponent || 'a';
-
-        return (
-          <td className={tdClassNames} style={containerStyle}>
-            <Tooltip
-              disabled={header.textWrapMethod !== 'truncate'}
-              enterDelay={tooltipDelay}
-              enterNextDelay={tooltipDelay}
-              placement={tooltipPlacement}
-              subtitle={content}
-            >
-              <LinkComponent
-                className={cellClassNames}
-                href={getRowLink(item)}
-                style={cellStyle}
-              >
-                {content}
-              </LinkComponent>
-            </Tooltip>
-          </td>
-        );
-      }
-
-      return (
-        <td className={tdClassNames} style={containerStyle}>
-          <Tooltip
-            disabled={header.textWrapMethod !== 'truncate'}
-            enterDelay={tooltipDelay}
-            enterNextDelay={tooltipDelay}
-            placement={tooltipPlacement}
-            subtitle={content}
-          >
-            {/* Child div required for proper truncating */}
-            <span className={cellClassNames} style={cellStyle}>
-              {content}
-            </span>
-          </Tooltip>
-        </td>
-      );
-    };
-
-    return header.renderCustom ? (
-      <>{renderContent(header.renderCustom(value, item))}</>
-    ) : (
-      <>{renderContent(String(value))}</>
-    );
-  };
-
   const renderRows = () => {
     if (loading) {
       return [...Array(loadingItems).keys()].map(n => (
@@ -344,114 +243,21 @@ export const SimpleTable = <T extends object>({
       ));
     }
 
-    if (collapsible.enabled && collapsible.collapseContent?.length) {
-      const { collapseContent, expandedItemKeys, toggleItem } = collapsible;
-      return items.map((item, index) => {
-        const key = keyExtractor(item);
-        const collapsed = !expandedItemKeys.includes(key);
-        const rowCollapseContent = collapseContent?.find(
-          x => x.key === key,
-        )?.content;
-        return (
-          <TableRowCollapse
-            key={key}
-            className={clsx(
-              !noHoverBackground && styles.withHover,
-              collapsed && styles.collapsed,
-              rowCollapseContent && styles.hasContent,
-            )}
-            collapsed={collapsed}
-            onToggleCollapse={() => {
-              toggleItem(key);
-              onRowClick?.(item);
-            }}
-            rowContent={
-              <>
-                {selectable.enabled && (
-                  <td>
-                    {selectable.renderCustomRowCheckbox?.(item, index) || (
-                      <Checkbox
-                        checked={
-                          selectable.isRowChecked?.(item, index) || false
-                        }
-                        disabled={
-                          selectable.isRowCheckboxDisabled?.(item, index) ||
-                          false
-                        }
-                        onChange={checked =>
-                          selectable.onRowCheckboxChange?.(checked, item, index)
-                        }
-                      />
-                    )}
-                  </td>
-                )}
-                {headers.map(header => (
-                  <Fragment key={header.key}>
-                    {renderHeader(header, item)}
-                  </Fragment>
-                ))}
-              </>
-            }
-          >
-            {rowCollapseContent}
-          </TableRowCollapse>
-        );
-      });
-    }
-
-    return items.map((item, index) => {
-      const checkboxRef = useRef<HTMLTableCellElement>(null);
-      const clickable =
-        !!onRowClick ||
-        (!!selectable.anySelected && !!selectable.onRowCheckboxChange);
-
-      return (
-        <tr
-          key={keyExtractor(item)}
-          className={clsx(
-            clickable && styles.clickable,
-            !noHoverBackground && styles.withHover,
-          )}
-          onClick={e => {
-            if (
-              selectable.anySelected ||
-              checkboxRef.current?.contains(e.target as Node)
-            ) {
-              if (!selectable.isRowCheckboxDisabled?.(item, index)) {
-                e.preventDefault();
-                selectable.onRowCheckboxChange?.(
-                  !selectable.isRowChecked?.(item, index),
-                  item,
-                  index,
-                );
-              }
-            } else {
-              onRowClick?.(item);
-            }
-          }}
-          onMouseEnter={() => onRowHover?.(item)}
-        >
-          {selectable.enabled && (
-            <td ref={checkboxRef}>
-              {selectable.renderCustomRowCheckbox?.(item, index) || (
-                <Checkbox
-                  checked={selectable.isRowChecked?.(item, index) || false}
-                  disabled={
-                    selectable.isRowCheckboxDisabled?.(item, index) || false
-                  }
-                  onChange={checked =>
-                    selectable.onRowCheckboxChange?.(checked, item, index)
-                  }
-                />
-              )}
-            </td>
-          )}
-          {headers.map(header => (
-            <Fragment key={header.key}>{renderHeader(header, item)}</Fragment>
-          ))}
-        </tr>
-      );
-    });
+    return items.map((item, index) => (
+      <SimpleTableRow
+        CustomLinkComponent={CustomLinkComponent}
+        collapsible={collapsible}
+        getRowLink={getRowLink}
+        headers={headers}
+        index={index}
+        item={item}
+        keyExtractor={keyExtractor}
+        noHoverBackground={noHoverBackground}
+        onRowClick={onRowClick}
+        onRowHover={onRowHover}
+        selectable={selectable}
+      />
+    ));
   };
 
   return (
